Move error formatters to lib
