+++ libs/errors/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/errors';
+export * from './lib/formatters';
